CIES/requirements.txt
beautifulsoup4==4.6.3
numpy==1.14.5
pandas==0.23.1
PyYAML==3.12
tqdm==4.26.0